refactor: extract some ComboBoxMixin logic to separate mixin
