+++ packages/combo-box/src/vaadin-combo-box-mixin.js
@@ -3,18 +3,10 @@
  * Copyright (c) 2015 - 2023 Vaadin Ltd.
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
-import { DisabledMixin } from '@vaadin/a11y-base/src/disabled-mixin.js';
-import { FocusMixin } from '@vaadin/a11y-base/src/focus-mixin.js';
-import { isElementFocused } from '@vaadin/a11y-base/src/focus-utils.js';
-import { KeyboardMixin } from '@vaadin/a11y-base/src/keyboard-mixin.js';
-import { isTouch } from '@vaadin/component-base/src/browser-utils.js';
-import { ControllerMixin } from '@vaadin/component-base/src/controller-mixin.js';
-import { OverlayClassMixin } from '@vaadin/component-base/src/overlay-class-mixin.js';
 import { get } from '@vaadin/component-base/src/path-utils.js';
 import { processTemplates } from '@vaadin/component-base/src/templates.js';
-import { InputMixin } from '@vaadin/field-base/src/input-mixin.js';
 import { ValidateMixin } from '@vaadin/field-base/src/validate-mixin.js';
-import { VirtualKeyboardController } from '@vaadin/field-base/src/virtual-keyboard-controller.js';
+import { ComboBoxBaseMixin } from './vaadin-combo-box-base-mixin.js';
 import { ComboBoxPlaceholder } from './vaadin-combo-box-placeholder.js';
 
 /**
@@ -47,51 +39,14 @@ function findItemIndex(items, callback) {
 
 /**
  * @polymerMixin
- * @mixes ControllerMixin
+ * @mixes ComboBoxBaseMixin
  * @mixes ValidateMixin
- * @mixes DisabledMixin
- * @mixes InputMixin
- * @mixes KeyboardMixin
- * @mixes FocusMixin
- * @mixes OverlayClassMixin
  * @param {function(new:HTMLElement)} subclass
  */
 export const ComboBoxMixin = (subclass) =>
-  class ComboBoxMixinClass extends OverlayClassMixin(
-    ControllerMixin(ValidateMixin(FocusMixin(KeyboardMixin(InputMixin(DisabledMixin(subclass)))))),
-  ) {
+  class ComboBoxMixinClass extends ComboBoxBaseMixin(ValidateMixin(subclass)) {
     static get properties() {
       return {
-        /**
-         * True if the dropdown is open, false otherwise.
-         * @type {boolean}
-         */
-        opened: {
-          type: Boolean,
-          notify: true,
-          value: false,
-          reflectToAttribute: true,
-          observer: '_openedChanged',
-        },
-
-        /**
-         * Set true to prevent the overlay from opening automatically.
-         * @attr {boolean} auto-open-disabled
-         */
-        autoOpenDisabled: {
-          type: Boolean,
-        },
-
-        /**
-         * When present, it specifies that the field is read-only.
-         * @type {boolean}
-         */
-        readonly: {
-          type: Boolean,
-          value: false,
-          reflectToAttribute: true,
-        },
-
         /**
          * Custom function for rendering the content of every item.
          * Receives three arguments:
@@ -226,15 +181,6 @@ export const ComboBoxMixin = (subclass) =>
         itemIdPath: String,
 
         /**
-         * @type {!HTMLElement | undefined}
-         * @protected
-         */
-        _toggleElement: {
-          type: Object,
-          observer: '_toggleElementChanged',
-        },
-
-        /**
          * Set of items to be rendered in the dropdown.
          * @protected
          */
@@ -244,9 +190,6 @@ export const ComboBoxMixin = (subclass) =>
 
         /** @private */
         _closeOnBlurIsPrevented: Boolean,
-
-        /** @private */
-        _scroller: Object,
 
         /** @private */
         _overlayOpened: {
@@ -267,29 +210,7 @@ export const ComboBoxMixin = (subclass) =>
     constructor() {
       super();
       this._boundOverlaySelectedItemChanged = this._overlaySelectedItemChanged.bind(this);
-      this._boundOnClearButtonMouseDown = this.__onClearButtonMouseDown.bind(this);
-      this._boundOnClick = this._onClick.bind(this);
       this._boundOnOverlayTouchAction = this._onOverlayTouchAction.bind(this);
-      this._boundOnTouchend = this._onTouchend.bind(this);
-    }
-
-    /**
-     * Tag name prefix used by scroller and items.
-     * @protected
-     * @return {string}
-     */
-    get _tagNamePrefix() {
-      return 'vaadin-combo-box';
-    }
-
-    /**
-     * Get a reference to the native `<input>` element.
-     * Override to provide a custom input.
-     * @protected
-     * @return {HTMLInputElement | undefined}
-     */
-    get _nativeInput() {
-      return this.inputElement;
     }
 
     /**
@@ -301,27 +222,8 @@ export const ComboBoxMixin = (subclass) =>
     _inputElementChanged(inputElement) {
       super._inputElementChanged(inputElement);
 
-      const input = this._nativeInput;
-
-      if (input) {
-        input.autocomplete = 'off';
-        input.autocapitalize = 'off';
-
-        input.setAttribute('role', 'combobox');
-        input.setAttribute('aria-autocomplete', 'list');
-        input.setAttribute('aria-expanded', !!this.opened);
-
-        // Disable the macOS Safari spell check auto corrections.
-        input.setAttribute('spellcheck', 'false');
-
-        // Disable iOS autocorrect suggestions.
-        input.setAttribute('autocorrect', 'off');
-
+      if (this._nativeInput) {
         this._revertInputValueToValue();
-
-        if (this.clearElement) {
-          this.clearElement.addEventListener('mousedown', this._boundOnClearButtonMouseDown);
-        }
       }
     }
 
@@ -329,34 +231,9 @@ export const ComboBoxMixin = (subclass) =>
     ready() {
       super.ready();
 
-      this._initOverlay();
-      this._initScroller();
-
       this._lastCommittedValue = this.value;
 
-      this.addEventListener('click', this._boundOnClick);
-      this.addEventListener('touchend', this._boundOnTouchend);
-
-      const bringToFrontListener = () => {
-        requestAnimationFrame(() => {
-          this._overlayElement.bringToFront();
-        });
-      };
-
-      this.addEventListener('mousedown', bringToFrontListener);
-      this.addEventListener('touchstart', bringToFrontListener);
-
       processTemplates(this);
-
-      this.addController(new VirtualKeyboardController(this));
-    }
-
-    /** @protected */
-    disconnectedCallback() {
-      super.disconnectedCallback();
-
-      // Close the overlay on detach
-      this.close();
     }
 
     /**
@@ -378,23 +255,6 @@ export const ComboBoxMixin = (subclass) =>
     }
 
     /**
-     * Opens the dropdown list.
-     */
-    open() {
-      // Prevent _open() being called when input is disabled or read-only
-      if (!this.disabled && !this.readonly) {
-        this.opened = true;
-      }
-    }
-
-    /**
-     * Closes the dropdown list.
-     */
-    close() {
-      this.opened = false;
-    }
-
-    /**
      * Override Polymer lifecycle callback to handle `filter` property change after
      * the observer for `opened` property is triggered. This is needed when opening
      * combo-box on user input to ensure the focused index is set correctly.
@@ -413,25 +273,19 @@ export const ComboBoxMixin = (subclass) =>
       }
     }
 
-    /** @private */
+    /** @protected */
     _initOverlay() {
-      const overlay = this.$.overlay;
+      super._initOverlay();
 
-      // Store instance for detecting "dir" attribute on opening
-      overlay._comboBox = this;
+      const overlay = this.$.overlay;
 
       overlay.addEventListener('touchend', this._boundOnOverlayTouchAction);
       overlay.addEventListener('touchmove', this._boundOnOverlayTouchAction);
-
-      // Prevent blurring the input when clicking inside the overlay
-      overlay.addEventListener('mousedown', (e) => e.preventDefault());
 
       // Manual two-way binding for the overlay "opened" property
       overlay.addEventListener('opened-changed', (e) => {
         this._overlayOpened = e.detail.value;
       });
-
-      this._overlayElement = overlay;
     }
 
     /**
@@ -439,29 +293,14 @@ export const ComboBoxMixin = (subclass) =>
      * Override to provide custom host reference.
      *
      * @protected
+     * @override
      */
     _initScroller(host) {
-      const scrollerTag = `${this._tagNamePrefix}-scroller`;
+      super._initScroller(host);
 
-      const overlay = this._overlayElement;
+      const scroller = this._scroller;
 
-      overlay.renderer = (root) => {
-        if (!root.firstChild) {
-          root.appendChild(document.createElement(scrollerTag));
-        }
-      };
-
-      // Ensure the scroller is rendered
-      overlay.requestContentUpdate();
-
-      const scroller = overlay.querySelector(scrollerTag);
-
-      scroller.owner = host || this;
-      scroller.getItemLabel = this._getItemLabel.bind(this);
       scroller.addEventListener('selection-changed', this._boundOverlaySelectedItemChanged);
-
-      // Trigger the observer to set properties
-      this._scroller = scroller;
     }
 
     /** @private */
@@ -514,11 +353,6 @@ export const ComboBoxMixin = (subclass) =>
       this._updateActiveDescendant(index);
     }
 
-    /** @protected */
-    _isInputFocused() {
-      return this.inputElement && isElementFocused(this.inputElement);
-    }
-
     /** @private */
     _updateActiveDescendant(index) {
       const input = this._nativeInput;
@@ -534,40 +368,15 @@ export const ComboBoxMixin = (subclass) =>
       }
     }
 
-    /** @private */
+    /**
+     * @protected
+     * @override
+     */
     _openedChanged(opened, wasOpened) {
-      // Prevent _close() being called when opened is set to its default value (false).
-      if (wasOpened === undefined) {
-        return;
-      }
+      super._openedChanged(opened, wasOpened);
 
-      if (opened) {
-        this._openedWithFocusRing = this.hasAttribute('focus-ring');
-        // For touch devices, we don't want to popup virtual keyboard
-        // unless input element is explicitly focused by the user.
-        if (!this._isInputFocused() && !isTouch) {
-          if (this.inputElement) {
-            this.inputElement.focus();
-          }
-        }
-
-        this._overlayElement.restoreFocusOnClose = true;
-      } else {
+      if (wasOpened) {
         this._onClosed();
-        if (this._openedWithFocusRing && this._isInputFocused()) {
-          this.setAttribute('focus-ring', '');
-        }
-      }
-
-      const input = this._nativeInput;
-      if (input) {
-        input.setAttribute('aria-expanded', !!opened);
-
-        if (opened) {
-          input.setAttribute('aria-controls', this._scroller.id);
-        } else {
-          input.removeAttribute('aria-controls');
-        }
       }
     }
 
@@ -580,66 +389,16 @@ export const ComboBoxMixin = (subclass) =>
       this._closeOnBlurIsPrevented = false;
     }
 
-    /** @protected */
-    _isClearButton(event) {
-      return event.composedPath()[0] === this.clearElement;
-    }
-
-    /** @private */
-    __onClearButtonMouseDown(event) {
-      event.preventDefault(); // Prevent native focusout event
-      this.inputElement.focus();
-    }
-
     /**
      * @param {Event} event
      * @protected
      */
     _onClearButtonClick(event) {
-      event.preventDefault();
-      this._onClearAction();
+      super._onClearButtonClick(event);
 
       // De-select dropdown item
       if (this.opened) {
         this.requestContentUpdate();
-      }
-    }
-
-    /**
-     * @param {Event} event
-     * @private
-     */
-    _onToggleButtonClick(event) {
-      // Prevent parent components such as `vaadin-grid`
-      // from handling the click event after it bubbles.
-      event.preventDefault();
-
-      if (this.opened) {
-        this.close();
-      } else {
-        this.open();
-      }
-    }
-
-    /**
-     * @param {Event} event
-     * @protected
-     */
-    _onHostClick(event) {
-      if (!this.autoOpenDisabled) {
-        event.preventDefault();
-        this.open();
-      }
-    }
-
-    /** @private */
-    _onClick(event) {
-      if (this._isClearButton(event)) {
-        this._onClearButtonClick(event);
-      } else if (event.composedPath().includes(this._toggleElement)) {
-        this._onToggleButtonClick(event);
-      } else {
-        this._onHostClick(event);
       }
     }
 
@@ -668,13 +427,12 @@ export const ComboBoxMixin = (subclass) =>
       }
     }
 
-    /** @private */
+    /**
+     * @protected
+     * @override
+     */
     _getItemLabel(item) {
-      let label = item && this.itemLabelPath ? get(this.itemLabelPath, item) : undefined;
-      if (label === undefined || label === null) {
-        label = item ? item.toString() : '';
-      }
-      return label;
+      return super._getItemLabel(item, this.itemLabelPath);
     }
 
     /** @private */
@@ -836,20 +594,6 @@ export const ComboBoxMixin = (subclass) =>
         e.stopPropagation();
         // The clear button is visible and the overlay is closed, so clear the value.
         this._onClearAction();
-      }
-    }
-
-    /** @private */
-    _toggleElementChanged(toggleElement) {
-      if (toggleElement) {
-        // Don't blur the input on toggle mousedown
-        toggleElement.addEventListener('mousedown', (e) => e.preventDefault());
-        // Unfocus previously focused element if focus is not inside combo box (on touch devices)
-        toggleElement.addEventListener('click', () => {
-          if (isTouch && !this._isInputFocused()) {
-            document.activeElement.blur();
-          }
-        });
       }
     }
 
@@ -1303,16 +1047,6 @@ export const ComboBoxMixin = (subclass) =>
       }
 
       return true;
-    }
-
-    /** @private */
-    _onTouchend(event) {
-      if (!this.clearElement || event.composedPath()[0] !== this.clearElement) {
-        return;
-      }
-
-      event.preventDefault();
-      this._onClearAction();
     }
 
     /**
